[Server] Move type-utils outside models folders
models/index.ts was trying to import this as a model and failing.
