+++ server/src/db/models/session.ts
@@ -1,5 +1,5 @@
 import { Sequelize, DataTypes, Model } from 'sequelize'
-import { ModelStatic } from './type-utils'
+import { ModelStatic } from '../type-utils'
 
 export interface Session extends Model {
   readonly sid: string
